split graphinputqueue impls into two nicely laid out files :D
